# gen/openapi/api/v1/auth_gateway_service.swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "api/v1/auth_gateway_service.proto",
    "version": "version not set"
  },
  "tags": [
    {
      "name": "AuthGatewayService"
    }
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/v1/auth/login": {
      "post": {
        "operationId": "AuthGatewayService_Login",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1AuthGatewayServiceLoginResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/googlerpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/v1AuthGatewayServiceLoginRequest"
            }
          }
        ],
        "tags": [
          "AuthGatewayService"
        ]
      }
    },
    "/v1/auth/logout": {
      "post": {
        "operationId": "AuthGatewayService_Logout",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1AuthGatewayServiceLogoutResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/googlerpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/v1AuthGatewayServiceLogoutRequest"
            }
          }
        ],
        "tags": [
          "AuthGatewayService"
        ]
      }
    },
    "/v1/auth/refresh": {
      "post": {
        "operationId": "AuthGatewayService_RefreshToken",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1AuthGatewayServiceRefreshTokenResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/googlerpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/v1AuthGatewayServiceRefreshTokenRequest"
            }
          }
        ],
        "tags": [
          "AuthGatewayService"
        ]
      }
    }
  },
  "definitions": {
    "commonv1Status": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "format": "int32",
          "description": "The status code, which should be an enum value of google.rpc.Code."
        },
        "message": {
          "type": "string",
          "description": "A developer-facing error message."
        },
        "details": {
          "type": "object",
          "description": "A list of messages that carry the error details."
        }
      },
      "description": "Status is a standard message for returning operation status.\nIt's modeled after google.rpc.Status."
    },
    "googlerpcStatus": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "format": "int32"
        },
        "message": {
          "type": "string"
        },
        "details": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/protobufAny"
          }
        }
      }
    },
    "protobufAny": {
      "type": "object",
      "properties": {
        "@type": {
          "type": "string"
        }
      },
      "additionalProperties": {}
    },
    "protobufNullValue": {
      "type": "string",
      "enum": [
        "NULL_VALUE"
      ],
      "default": "NULL_VALUE",
      "description": "`NullValue` is a singleton enumeration to represent the null value for the\n`Value` type union.\n\nThe JSON representation for `NullValue` is JSON `null`.\n\n - NULL_VALUE: Null value."
    },
    "v1AuthGatewayServiceLoginRequest": {
      "type": "object",
      "properties": {
        "loginRequest": {
          "$ref": "#/definitions/v1LoginRequest"
        }
      }
    },
    "v1AuthGatewayServiceLoginResponse": {
      "type": "object",
      "properties": {
        "loginResponse": {
          "$ref": "#/definitions/v1LoginResponse"
        }
      }
    },
    "v1AuthGatewayServiceLogoutRequest": {
      "type": "object",
      "properties": {
        "empty": {
          "type": "object",
          "properties": {}
        }
      }
    },
    "v1AuthGatewayServiceLogoutResponse": {
      "type": "object",
      "properties": {
        "logoutResponse": {
          "$ref": "#/definitions/v1LogoutResponse"
        }
      }
    },
    "v1AuthGatewayServiceRefreshTokenRequest": {
      "type": "object",
      "properties": {
        "refreshTokenRequest": {
          "$ref": "#/definitions/v1RefreshTokenRequest"
        }
      }
    },
    "v1AuthGatewayServiceRefreshTokenResponse": {
      "type": "object",
      "properties": {
        "refreshTokenResponse": {
          "$ref": "#/definitions/v1RefreshTokenResponse"
        }
      }
    },
    "v1AuthToken": {
      "type": "object",
      "properties": {
        "accessToken": {
          "type": "string",
          "description": "The access token for API requests."
        },
        "refreshToken": {
          "type": "string",
          "description": "The refresh token to obtain a new access token."
        },
        "expiresAt": {
          "type": "string",
          "format": "date-time",
          "description": "The expiration time of the access token."
        },
        "tokenType": {
          "$ref": "#/definitions/v1TokenType",
          "description": "The type of the token, typically \"Bearer\"."
        },
        "sessionId": {
          "type": "string",
          "description": "The unique identifier for the session."
        }
      },
      "description": "AuthToken represents the tokens issued upon successful authentication."
    },
    "v1LoginRequest": {
      "type": "object",
      "properties": {
        "email": {
          "type": "string"
        },
        "password": {
          "type": "string"
        }
      }
    },
    "v1LoginResponse": {
      "type": "object",
      "properties": {
        "status": {
          "$ref": "#/definitions/commonv1Status"
        },
        "token": {
          "$ref": "#/definitions/v1AuthToken"
        },
        "user": {
          "$ref": "#/definitions/v1UserProfile"
        }
      }
    },
    "v1LogoutResponse": {
      "type": "object",
      "properties": {
        "status": {
          "$ref": "#/definitions/commonv1Status"
        }
      }
    },
    "v1RefreshTokenRequest": {
      "type": "object",
      "properties": {
        "refreshToken": {
          "type": "string"
        }
      }
    },
    "v1RefreshTokenResponse": {
      "type": "object",
      "properties": {
        "status": {
          "$ref": "#/definitions/commonv1Status"
        },
        "token": {
          "$ref": "#/definitions/v1AuthToken"
        }
      }
    },
    "v1TokenType": {
      "type": "string",
      "enum": [
        "TOKEN_TYPE_UNSPECIFIED",
        "TOKEN_TYPE_BEARER"
      ],
      "default": "TOKEN_TYPE_UNSPECIFIED",
      "description": "The type of the authentication token.\n\n - TOKEN_TYPE_BEARER: Standard bearer token."
    },
    "v1UserProfile": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string"
        },
        "email": {
          "type": "string"
        },
        "username": {
          "type": "string"
        },
        "firstName": {
          "type": "string"
        },
        "lastName": {
          "type": "string"
        },
        "profileImageUrl": {
          "type": "string"
        },
        "createdAt": {
          "type": "string",
          "format": "date-time"
        },
        "updatedAt": {
          "type": "string",
          "format": "date-time"
        }
      },
      "description": "UserProfile contains public information about a user."
    }
  }
}